UPD: review code from command module
