+++ Makefile
@@ -2,7 +2,7 @@ all: test
 
 test:
 	@echo "Running tests..."
-	pytest -v --cov ./unfazed/cache/ --cov-report term-missing
+	pytest -v --cov ./unfazed/command/ --cov-report term-missing
 
 format:
 	@echo "Formatting code..."
