--- tests/apps/cmd/wrong/app.py
@@ -1,6 +1,0 @@
-from unfazed.app import BaseAppConfig
-
-
-class AppConfig(BaseAppConfig):
-    async def ready(self) -> None:
-        print("AppConfig is ready!")
