--- tests/apps/cmd/wrong/commands/noasync.py
@@ -1,8 +1,0 @@
-import typing as t
-
-from unfazed.command import BaseCommand
-
-
-class Command(BaseCommand):
-    async def handle(self, **options: t.Any) -> None:
-        print("Command is handled!")
