--- tests/apps/cmd/wrong/commands/nohandle.py
@@ -1,5 +1,0 @@
-from unfazed.command import BaseCommand
-
-
-class Command(BaseCommand):
-    pass
