+++ tests/test_command/test_internal_commands/test_start_cmd.py
